requirements.txt
click==8.1.3
easydict==1.9
matplotlib==3.5.1
MinkowskiEngine==0.5.4
numpy==1.20.3
open3d==0.13.0
pykeops==2.1
pytorch_lightning==1.5.6
PyYAML==6.0
scipy==1.7.3
torch_scatter==2.0.9
torch_geometric==1.7.2
spconv-cu113==2.1.21
torch_sparse==0.6.12
cumm-cu113==0.2.8 
torch_cluster==1.5.9
tensorboard
timm
termcolor
tensorboardX